webapp/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.3
opencv-python==4.9.0.80
pillow==10.4.0
numpy==2.2.6
jinja2==3.1.4
python-multipart==0.0.9
